# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: ActionLoom AI Agent API
  description: |
    ActionLoom AI Agent API provides programmatic access to blockchain workflow composition, 
    Action discovery, and Agent management capabilities. This API is designed for AI agents, 
    trading bots, and automated systems to integrate with Flow blockchain automation.
    
    ## Authentication
    All API endpoints require authentication using API keys. Include your API key in the 
    Authorization header as a Bearer token:
    
    ```
    Authorization: Bearer al_your_api_key_here
    ```
    
    ## Rate Limits
    API requests are rate-limited based on your subscription tier:
    
    - **Free Tier**: 10 requests/minute, 100 requests/hour, 1,000 requests/day
    - **Pro Tier**: 100 requests/minute, 5,000 requests/hour, 50,000 requests/day  
    - **Enterprise Tier**: 1,000 requests/minute, 50,000 requests/hour, 1,000,000 requests/day
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Remaining-Minute`: Remaining requests this minute
    - `X-RateLimit-Remaining-Hour`: Remaining requests this hour
    - `X-RateLimit-Remaining-Day`: Remaining requests today
    
    ## Error Handling
    The API uses standard HTTP status codes and returns consistent error responses:
    
    ```json
    {
      "success": false,
      "error": "Error description",
      "details": "Additional error details"
    }
    ```
    
    Common status codes:
    - `200`: Success
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /actions:
    get:
      summary: Discover Actions
      description: |
        Discover and search available blockchain Actions from Flow's on-chain registries.
        Supports filtering by category, searching by query, and retrieving specific Actions by ID.
      operationId: discoverActions
      tags:
        - Action Discovery
      parameters:
        - name: q
          in: query
          description: Search query for Actions
          schema:
            type: string
          example: "swap"
        - name: category
          in: query
          description: Filter Actions by category
          schema:
            type: string
          example: "defi"
        - name: id
          in: query
          description: Get specific Action by ID
          schema:
            type: string
          example: "action_swap_usdc_flow"
        - name: limit
          in: query
          description: Maximum number of Actions to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: refresh
          in: query
          description: Force refresh of cached Action data
          schema:
            type: boolean
            default: false
        - name: similar
          in: query
          description: Find Actions similar to the specified Action ID
          schema:
            type: string
      responses:
        '200':
          description: Actions discovered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsResponse'
        '404':
          description: Action not found (when requesting specific Action by ID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Validate Actions
      description: |
        Validate Action metadata, check compatibility between Actions, or validate 
        workflow chains of Actions.
      operationId: validateActions
      tags:
        - Action Discovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ValidateActionRequest'
                - $ref: '#/components/schemas/ValidateCompatibilityRequest'
                - $ref: '#/components/schemas/ValidateChainRequest'
            examples:
              validateAction:
                summary: Validate Single Action
                value:
                  action:
                    id: "swap_action"
                    name: "Token Swap"
                    description: "Swap tokens on DEX"
                    category: "defi"
                    inputs:
                      - name: "fromToken"
                        type: "string"
                        required: true
                    outputs:
                      - name: "toToken"
                        type: "string"
              validateCompatibility:
                summary: Check Action Compatibility
                value:
                  sourceActionId: "action1"
                  targetActionId: "action2"
              validateChain:
                summary: Validate Workflow Chain
                value:
                  actionIds: ["action1", "action2", "action3"]
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key authentication. Include your API key in the Authorization header:
        `Authorization: Bearer al_your_api_key_here`
  schemas:
...
    ActionsResponse:
      type: object
      properties:
        success:
          type: boolean
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionMetadata'
        query:
          type: string
          description: Search query used (if applicable)
        category:
          type: string
          description: Category filter used (if applicable)
        total:
          type: integer
          description: Total number of actions found
        registries:
          type: array
          items:
            $ref: '#/components/schemas/ActionRegistry'
        lastUpdated:
          type: string
          format: date-time
        executionTime:
          type: integer
          description: Discovery execution time in milliseconds
    ActionMetadata:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        version:
          type: string
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/ActionInput'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/ActionOutput'
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/ActionParameter'
        compatibility:
          $ref: '#/components/schemas/CompatibilityInfo'
        gasEstimate:
          type: integer
        securityLevel:
          type: string
          enum: [low, medium, high, critical]
        author:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ActionInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        required:
          type: boolean
        description:
          type: string
        validation:
          type: array
          items:
            $ref: '#/components/schemas/ValidationRule'
    ActionOutput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
    ActionParameter:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        required:
          type: boolean
    ValidationRule:
      type: object
      properties:
        type:
          type: string
          enum: [range, pattern, custom]
        value:
          oneOf:
            - type: string
            - type: number
            - type: object
        message:
          type: string
    CompatibilityInfo:
      type: object
      properties:
        requiredCapabilities:
          type: array
          items:
            type: string
        supportedNetworks:
          type: array
          items:
            type: string
        minimumFlowVersion:
          type: string
        conflictsWith:
          type: array
          items:
            type: string
    ActionRegistry:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        description:
          type: string
        actions:
          type: array
          items:
            type: string

    # Validation schemas
    ValidateActionRequest:
      type: object
      required: [action]
      properties:
        action:
          $ref: '#/components/schemas/ActionMetadata'
    ValidateCompatibilityRequest:
      type: object
      required: [sourceActionId, targetActionId]
      properties:
        sourceActionId:
          type: string
        targetActionId:
          type: string
    ValidateChainRequest:
      type: object
      required: [actionIds]
      properties:
        actionIds:
          type: array
          items:
            type: string
    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
        validationResult:
          $ref: '#/components/schemas/ValidationResult'
        compatible:
          type: boolean
        issues:
          type: array
          items:
            $ref: '#/components/schemas/CompatibilityIssue'
        chainValidation:
          $ref: '#/components/schemas/ValidationResult'
    ValidationResult:
      type: object
      properties:
        isValid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        warnings:
          type: array
          items:
            type: string
        compatibilityIssues:
          type: array
          items:
            $ref: '#/components/schemas/CompatibilityIssue'
    ValidationError:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
        actionId:
          type: string
        field:
          type: string
        severity:
          type: string
          enum: [error, warning]
    CompatibilityIssue:
      type: object
      properties:
        sourceActionId:
          type: string
        targetActionId:
          type: string
        issue:
          type: string
        suggestion:
          type: string

    # Agent-related schemas
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional technical details about the error
        code:
          type: string
          description: Machine-readable error code
...